+++ hyperSpec/R/wl_eval.R
@@ -1,24 +1,40 @@
-#' Evaluate function on wavelengths of `hyperSpec` object.
+#' Evaluate function on wavelengths of `hyperSpec` object
 #'
 #' This is useful for generating certain types of baseline "reference spectra".
 #'
-#' @param x `hyperSpec` object
-#' @param ... hyperSpec method: expressions to be evaluated
-#' @param normalize.wl function to transorm the wavelengths before evaluating the polynomial (or
-#' other function). Use [hyperSpec::normalize01()] to map the wavelength range to the interval \[0, 1\].
-#' @return `hyperSpec` object containing one spectrum for each expression
+#' @param x either `hyperSpec` object or numeric vector.
+#' @param ... expressions to be evaluated.
+#' @param normalize.wl function to transorm the wavelengths before evaluating
+#' the polynomial (or  other function). Use [hyperSpec::normalize01()] to map
+#' the wavelength range to the interval \[0, 1\].
+#' @return `hyperSpec` object containing one spectrum for each expression.
 #'
 #' @export
 #'
 #' @concept wavelengths
 #'
-#' @seealso [hyperSpec::vanderMonde()] for  polynomials,
+#' @seealso
 #'
-#' [hyperSpec::normalize01()] to normalize the wavenumbers before evaluating the function
-#' @author C. Beleites
+#' - [hyperSpec::vanderMonde()] for  polynomials,
+#' - [hyperSpec::normalize01()] to normalize the wavenumbers before evaluating
+#' the function.
+#'
+#' @author C. Beleites, V. Gegzna
+#'
 #' @examples
 #' plot(wl.eval(laser, exp = function(x) exp(-x)))
+#'
+#' plot(wl.eval(1000:4000, y = function(x) 1/log(x)))
+#'
+#' plot(wl.eval(300:550, y2 = function(x) x*2, y3 = function(x) x*3))
+#'
 wl.eval <- function(x, ..., normalize.wl = I) {
+  UseMethod("wl.eval")
+}
+
+#' @rdname wl.eval
+#' @export
+wl.eval.hyperSpec <- function(x, ..., normalize.wl = I) {
   chk.hy(x)
   validObject(x)
 
@@ -36,14 +52,29 @@ wl.eval <- function(x, ..., normalize.wl = I) {
 }
 
 
-hySpc.testthat::test(wl.eval) <- function() {
+#' @rdname wl.eval
+#' @export
+wl.eval.numeric <- function(x, ..., normalize.wl = I) {
+  if (!is.vector(x)) {
+    class_txt <- paste(class(x), collapse = ", ")
+    stop("`x` must be a vector. Now it is ", class_txt, ".")
+  }
+  x <- new("hyperSpec", spc = seq_along(x), wavelength = x)
+  wl.eval(x, ..., normalize.wl = normalize.wl)
+}
+
+
+# Unit tests -----------------------------------------------------------------
+
+
+hySpc.testthat::test(wl.eval.hyperSpec) <- function() {
   context("wl.eval")
 
   test_that("error on function not returning same length as input", {
     expect_error(wl.eval(flu, function(x) 1))
   })
 
-  test_that("wl.eval against manual evaluation", {
+  test_that("wl.eval(<hyperSpec>) against manual evaluation", {
     expect_equivalent(
       wl.eval(flu, function(x) rep(5, length(x)), normalize.wl = normalize01)[[]],
       matrix(rep(5, nwl(flu)), nrow = 1)
@@ -89,5 +120,33 @@ hySpc.testthat::test(wl.eval) <- function() {
     tmp <- wl.eval(flu, f = function(x) x, g = function(x) exp(-x))
 
     expect_equal(tmp$.f, c("f", "g"))
+  })
+
+  test_that("wl.eval(<numeric>) works", {
+
+    expect_equal(
+      as.vector(wl.eval(1:10, f = function(x) x)$spc),
+      1:10
+    )
+
+    expect_equal(
+      as.vector(wl.eval(1:10, f = function(x) x**2)$spc),
+      (1:10)**2
+    )
+
+    expect_equal(
+      wl.eval(wl(flu), f = function(x) x)$.f,
+      wl.eval(   flu,  f = function(x) x)$.f
+    )
+
+    expect_silent(
+      tmp <- wl.eval(300:500, f = function(x) x, g = function(x) exp(-x))
+    )
+    expect_equal(tmp$.f, c("f", "g"))
+  })
+
+  test_that("wl.eval fails with matrix input", {
+    expect_error(wl.eval(matrix(1:10), f = function(x) x))
+    expect_error(wl.eval(matrix(), f = function(x) x))
   })
 }